+++ CHANGELOG.md
@@ -1,3 +1,19 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* FEATURE : SocketIO android work.
+
+## 0.0.1+1
+
+* CHANGE: Add licence
+
+## 0.0.1+2
+
+* CHANGE: Add description
+
+## 0.0.1+3
+
+* CHANGE: Update description
+
+## 0.0.1+4
+
+* Activate autoConnect